requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
addict=2.4.0=pypi_0
anyio=3.6.1=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
attrs=22.1.0=pypi_0
autopep8=1.6.0=pyhd3eb1b0_0
babel=2.10.3=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.4.26=h06a4308_0
certifi=2022.5.18.1=py38h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.0=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
defusedxml=0.7.1=pypi_0
deprecation=2.1.0=pypi_0
entrypoints=0.4=pypi_0
expat=2.4.4=h295c915_0
fastjsonschema=2.16.1=pypi_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fvcore=0.1.5.post20210915=py38
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.3=pypi_0
importlib-resources=5.9.0=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
iopath=0.1.9=py38
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
json5=0.9.9=pypi_0
jsonschema=4.9.1=pypi_0
jupyter-packaging=0.12.2=pypi_0
jupyter-server=1.18.1=pypi_0
jupyterlab=3.4.5=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.15.0=pypi_0
jupyterlab-widgets=1.1.1=pypi_0
kiwisolver=1.4.2=py38h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.12=h03d6c58_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=py38h06a4308_1
matplotlib-base=3.5.1=py38ha18d171_1
mistune=0.8.4=pypi_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=py_0
nbclassic=0.4.3=pypi_0
nbclient=0.6.6=pypi_0
nbconvert=6.5.3=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
notebook=6.4.12=pypi_0
notebook-shim=0.1.0=pypi_0
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
nvidiacub=1.10.0=0
olefile=0.46=py_0
open3d=0.15.2=pypi_0
openh264=2.1.0=hd408876_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1o=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.3=pypi_0
pandocfilters=1.5.0=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pypi_0
pillow=8.3.1=py38h2c7a002_0
pip=21.0.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
portalocker=2.4.0=py38h578d9bd_0
prometheus-client=0.14.1=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pypi_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pypi_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.9.2=py38h05f1152_4
pyquaternion=0.9.9=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.9.0=py3.8_cuda11.1_cudnn8.0.5_0
pytorch3d=0.6.2=py38_cu111_pyt190
pytz=2022.2.1=pypi_0
pyyaml=5.4.1=py38h497a2fe_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.28.1=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=65.0.0=pypi_0
sip=4.19.13=py38h295c915_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.36.0=hc218d9a_0
tabulate=0.8.9=pyhd8ed1ab_0
termcolor=1.1.0=py_2
terminado=0.15.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pyhd3eb1b0_0
tomlkit=0.11.4=pypi_0
torchaudio=0.9.0=py38
torchmetrics=0.8.2=pyhd8ed1ab_0
torchvision=0.10.0=py38_cu111
tornado=6.1=py38h27cfd23_0
tqdm=4.63.0=pyhd3eb1b0_0
traitlets=5.3.0=pypi_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.11=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.3.3=pypi_0
wheel=0.37.0=pyhd3eb1b0_0
widgetsnbextension=3.6.1=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h516909a_0
zipp=3.8.1=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
